requirements.txt
# --- Flask core ---
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
python-dotenv==1.0.1

# --- Auth & OAuth ---
Authlib==1.6.5
Flask-Dance==7.1.0
cryptography==46.0.2
requests==2.32.5
google-auth==2.40.3
google-auth-oauthlib==1.2.1
google-api-python-client==2.182.0
supabase==2.21.1

# --- AI / Whisper / LangChain ---
torch==2.8.0
openai-whisper==20250625
gTTS==2.5.4
langchain==0.3.27
langchain-community==0.3.29
langchain-google-genai==2.0.10
duckduckgo-search
ddgs

# --- Utilities ---
PyPDF2==3.0.1
tqdm==4.67.1
pytz==2025.2